Big re-factor of monitors and report HTML related classes
