+++ pyworkflow/em/protocol/monitors/protocol_monitor_ctf.py
