+++ pyworkflow/em/protocol/monitors/protocol_monitor_system.py
