+++ pyworkflow/em/viewer_monitor_summary.py
@@ -24,24 +24,16 @@
 # *
 # **************************************************************************
 
-import os
 import Tkinter as tk
 
-from os.path import join
 
-import pyworkflow.object as pwobj
-import pyworkflow.utils as pwutils
-from pyworkflow.gui.tree import TreeProvider, BoundTree
-from pyworkflow.gui.widgets import Button, HotButton
-import pyworkflow.gui.text as text
 import pyworkflow.gui as pwgui
-
-import pyworkflow.protocol.params as params
+import pyworkflow.gui.text as text
+import pyworkflow.utils as pwutils
+from protocol.monitors import ProtMonitorSummary, SummaryProvider
+from pyworkflow.gui.tree import BoundTree
+from pyworkflow.gui.widgets import Button, HotButton
 from pyworkflow.viewer import DESKTOP_TKINTER, Viewer
-
-from protocol.protocol_import import ProtImportImages
-from protocol.protocol_monitor_summary import ProtMonitorSummary
-
 
 
 class ViewerMonitorSummary(Viewer):
@@ -57,69 +49,7 @@ class ViewerMonitorSummary(Viewer):
         self.summaryWindow.show()
 
 
-class SummaryProvider(TreeProvider):
-    """Create the tree elements for a Protocol run"""
-    def __init__(self, protocol):
-        self.protocol = protocol
-        self.getColumns = lambda: [('Name', 300), ('Output', 150),
-                                   ('Number', 100)]
-        self._parentDict = {}
-        self.acquisition = []
-        self.refreshObjects()
 
-    def getObjects(self):
-        return self._objects
-
-    def refreshObjects(self):
-        objects = []
-
-        def addObj(objId, name, output='', size='', parent=None):
-            obj = pwobj.Object(objId=objId)
-            obj.name = name
-            obj.output = output
-            obj.outSize = size
-            obj._objParent = parent
-            objects.append(obj)
-            return obj
-
-        runs = [p.get() for p in self.protocol.inputProtocols]
-        g = self.protocol.getProject().getGraphFromRuns(runs)
-
-        nodes = g.getRoot().iterChildsBreadth()
-
-        for n in nodes:
-            prot = n.run
-            pobj = addObj(prot.getObjId(),
-                          '%s (id=%s)' % (prot.getRunName(), prot.strId()))
-
-            for outName, outSet in prot.iterOutputAttributes(pwobj.Set):
-                outSet.load()
-                outSet.loadAllProperties()
-                addObj(outSet.getObjId(), '', outName, outSet.getSize(), pobj)
-                outSet.close()
-                # Store acquisition parameters in case of the import protocol
-                if isinstance(prot, ProtImportImages):
-                    self.acquisition = [("Microscope Voltage: ",
-                                         prot.voltage.get()),
-                                        ("Spherical aberration: ",
-                                         prot.sphericalAberration.get()),
-                                        ("Magnification: ",
-                                         prot.magnification.get()),
-                                        ("Pixel Size (A/px): ",
-                                         outSet.getSamplingRate())
-                                        ]
-
-        self._objects = objects
-
-    def getObjectInfo(self, obj):
-        info = {'key': obj.strId(),
-                'parent': obj._objParent,
-                'text': obj.name,
-                'values': (obj.output, obj.outSize),
-                'open': True
-               }
-
-        return info
 
 
 class SummaryWindow(pwgui.Window):
@@ -135,33 +65,6 @@ class SummaryWindow(pwgui.Window):
         self._createContent(content)
         content.grid(row=0, column=0, sticky='news')
         content.columnconfigure(0, weight=1)
-
-    @staticmethod
-    def getHTMLReportText():
-
-        htmlTemplate = SummaryWindow.getSummaryHTMLTemplate()
-
-        if htmlTemplate is not None:
-            return open(htmlTemplate, 'r').read()
-        else:
-            return ""
-        return
-
-    @staticmethod
-    def getSummaryHTMLTemplate():
-
-        htmlFile = SummaryWindow.getSummaryHTMLTemplatePath()
-
-        # If the html template does not exist..
-        if not (os.path.exists(htmlFile)):
-            print "Html template files not found. Please get a valid HTML template and put it at %s", htmlFile
-        else:
-            return htmlFile
-
-    @staticmethod
-    def getSummaryHTMLTemplatePath():
-
-        return join(pwutils.getTemplatesFolder(), 'execution.summary.template.html')
 
     def _createContent(self, content):
         topFrame = tk.Frame(content)
@@ -224,23 +127,19 @@ class SummaryWindow(pwgui.Window):
         sysBtn = Button(subframe, "System Monitor", command=self._monitorSystem)
         sysBtn.grid(row=0, column=1, sticky='nw', padx=(0, 5))
 
-        pdfBtn = HotButton(subframe, 'Generate PDF Report',
-                           command=self._generatePDF)
-        pdfBtn.grid(row=0, column=2, sticky='nw', padx=(0, 5))
-
         htmlBtn = HotButton(subframe, 'Generate HTML Report',
                            command=self._generateHTML)
-        htmlBtn.grid(row=0, column=3, sticky='nw', padx=(0, 5))
+        htmlBtn.grid(row=0, column=2, sticky='nw', padx=(0, 5))
 
         closeBtn = self.createCloseButton(frame)
         closeBtn.grid(row=0, column=1, sticky='ne')
 
     def _monitorCTF(self, e=None):
-        from pyworkflow.em.protocol.protocol_monitor_ctf import CtfMonitorPlotter
+        from pyworkflow.em.protocol.monitors import CtfMonitorPlotter
         CtfMonitorPlotter(self.protocol.createCtfMonitor()).show()
 
     def _monitorSystem(self, e=None):
-        from pyworkflow.em.protocol.protocol_monitor_system import SystemMonitorPlotter
+        from pyworkflow.em.protocol.monitors import SystemMonitorPlotter
         SystemMonitorPlotter(self.protocol.createSystemMonitor()).show()
 
     def _updateLabel(self):
@@ -253,134 +152,7 @@ class SummaryWindow(pwgui.Window):
         self.tree.update()
         self._updateLabel()
 
-    def _generatePDF(self, e=None):
-        reportName = 'report_%s.tex' % pwutils.prettyTimestamp()
-        reportPath = self.protocol._getExtraPath(reportName)
-
-        acquisitionLines = ''
-        for item in self.provider.acquisition:
-            acquisitionLines += '%s & %s \\\\\n' % item
-
-        runLines = ''
-        for obj in self.provider.getObjects():
-            if obj.name:
-                runLines += ' %s &  &  \\\\\n' % obj.name
-            else:
-                runLines += ' & %s & %s \\\\\n' % (obj.output, obj.outSize)
-
-        args = {'acquisitionLines': acquisitionLines,
-                'runLines': runLines,
-                'dateStr': pwutils.prettyTime(secs=True),
-                'projectName': self.protocol.getProject().getShortName(),
-                'scipionVersion': os.environ['SCIPION_VERSION']
-                }
-
-        reportTemplate = """
-\documentclass{article}
-
-\usepackage[table]{xcolor}
-\setlength{\\tabcolsep}{15pt}
-\\renewcommand{\\arraystretch}{1.5}
-\\begin{document}
-
-%% Heading
-\hfil{\Huge\\bf Scipion Report}\hfil
-\hfill \\break
-\hrule
-
-\\begin{itemize}
-\item Date: \\textbf{%(dateStr)s}
-\item Project: \\textbf{%(projectName)s}
-\item Scipion version: \\textbf{%(scipionVersion)s}
-\end{itemize}
-
-\hfill \\break
-\hfill \\break
-
-\\begin{tabular}{ |p{5cm} p{5cm}|  }
-\hline
-\multicolumn{2}{|c|}{\\textbf{Acquisition}} \\\\
-\hline
-%(acquisitionLines)s
-\hline
-\end{tabular}
-
-\hfill \\break
-\hfill \\break
-\hfill \\break
-
-\\begin{tabular}{ |p{5cm} p{3cm} p{3cm}|  }
-\hline
-\multicolumn{3}{|c|}{\\textbf{Runs Summary}} \\\\
-\hline
-%(runLines)s
-\hline
-\end{tabular}
-
-\end{document}
-
-        """
-        reportFile = open(reportPath, 'w')
-        reportFile.write(reportTemplate % args)
-        reportFile.close()
-        pwutils.runJob(None, 'pdflatex',
-                       '-interaction=nonstopmode ' + reportName,
-                       cwd=self.protocol._getExtraPath())
-        text._open_cmd(reportPath.replace('.tex', '.pdf'))
-
     def _generateHTML(self, e=None):
-        reportName = 'report_%s.html' % pwutils.prettyTimestamp()
-        reportPath = self.protocol._getExtraPath(reportName)
-
-        acquisitionLines = ''
-        for item in self.provider.acquisition:
-            if not acquisitionLines == '':
-                acquisitionLines += ','
-
-            acquisitionLines += '{propertyName:"%s", propertyValue:"%s"}' % item
-
-        runLines = ''
-        wasProtocol = None
-        for obj in self.provider.getObjects():
-
-            # If it's a protocol
-            isProtocol = True if obj.name else False
-
-            if isProtocol:
-                if runLines != '': runLines += ']},'
-                runLines += '{protocolName: "%s", output:[' % obj.name
-            else:
-                if not wasProtocol: runLines += ','
-                runLines += '{name: "%s",  size:"%s"}' % (obj.output, obj.outSize)
-
-            wasProtocol = isProtocol
-
-        # End the runLines JSON object
-        runLines += ']}'
-
-        # Charts
-        import json
-
-        # Ctf monitor chart data
-        data = self.protocol.createCtfMonitor().getData()
-        ctfData = json.dumps(data)
-
-        # system monitor chart data
-        data = self.protocol.createSystemMonitor().getData()
-        systemData = json.dumps(data)
-
-        args = {'acquisitionLines': acquisitionLines,
-                'runLines': runLines,
-                'dateStr': pwutils.prettyTime(secs=True),
-                'projectName': self.protocol.getProject().getShortName(),
-                'scipionVersion': os.environ['SCIPION_VERSION'],
-                'ctfData': ctfData,
-                'systemData': systemData
-                }
-
-        reportTemplate = SummaryWindow.getHTMLReportText()
-
-        reportFile = open(reportPath, 'w')
-        reportFile.write(reportTemplate % args)
-        reportFile.close()
+        reportHtml = self.protocol.createHtmlReport()
+        reportPath = reportHtml.generate()
         text._open_cmd(reportPath)
